+++ web/main.tsx
@@ -4,13 +4,19 @@ import { BrowserRouter } from 'react-router-dom'
 
 import { App } from './App'
 import './main.css'
+import { ApiProvider } from './lib/api/ApiProvider'
+import { ApiClient, DEFAULT_API_ENDPOINT } from './lib/api/api-client'
+
+const apiClient = new ApiClient(DEFAULT_API_ENDPOINT)
 
 const root = document.getElementById('root')
 if (root) {
   createRoot(root).render(
     <React.StrictMode>
       <BrowserRouter>
-        <App />
+        <ApiProvider client={apiClient}>
+          <App />
+        </ApiProvider>
       </BrowserRouter>
     </React.StrictMode>
   )
